requirement.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.1
boto3==1.26.44
botocore==1.29.44
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
decorator==5.1.1
Django==4.1.5
django-phone-field==1.8.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
dynaconf==3.1.11
et-xmlfile==1.1.0
hvac==1.0.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
pyhcl==0.4.4
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.1
requests==2.28.1
s3transfer==0.6.0
shortuuid==1.0.11
six==1.16.0
slack-bolt==1.16.1
slack-sdk==3.19.5
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
validators==0.20.0
win32-setctime==1.1.0
